feat: major contract refactor + multi-chain deployment + API sync
Contracts:
- AlphaToken → WorknetToken: non-upgradeable, OZ ERC1363/ERC20Permit,
  callback-based constructor (universal vanity salt), single-minter
  with time-based cap, no admin field
- AlphaTokenFactory → WorknetTokenFactory: pending params pattern,
  predictDeployAddress(salt) universal
- StakeNFT → veAWP, StakingVault → AWPAllocator, WorknetNFT → AWPWorkNet
- WorknetManager refactored: Base/PCS/Uni separation, LP position reuse,
  MEV-protected buyback (minAmountOut), storage packing, liquidity=0 guard
- AWPDAO: timestamp clock, quorum on raw staked amount (not time-weighted),
  proposalThreshold on raw staked amount (200K AWP), _proposalQuorumVotes
- OZ upgraded 5.1.0 → 5.6.1: ReentrancyGuardTransient (TSTORE ~5K gas/call),
  removed __UUPSUpgradeable_init, _castVote → _tallyUpdated
- LPManager: Permit2 expiration optimized, alphaAmount → worknetTokenAmount
- AWPEmission: OZ IERC1363Receiver, ReentrancyGuardTransient

Multi-chain deployment (Base/ETH/ARB/BSC):
- AWPWorkNet proxy: 0x00000bfb...00A7
- WorknetTokenFactory: 0x000058EF...00AF
- veAWP: 0x0000b534...00A8
- AWPAllocator proxy: 0x0000D6BB...00AA
- AWPDAO proxy: 0x00006879...0DA0
- AWPRegistry impl: 0x00001a5a...001a
- AWPEmission impl: 0x0000dd24...00ae
- WorknetManager impl: per-chain (0x0000???00A9)
- LPManager impl: per-chain (0x0000???00A5)
- All verified on Etherscan, Safe proposals submitted

API:
- All Go bindings regenerated (abigen) for renamed contracts
- Config env vars renamed (VEAWP_ADDRESS, AWP_ALLOCATOR_ADDRESS, etc.)
- DB schema columns renamed (worknet_token, veawp, awp_allocator, awp_worknet)
- Indexer event parsing updated for new ABI
- Handler JSON fields, Redis keys, API methods updated
- Zero remaining old name references

Tests:
- 268 unit/integration tests passing
- 58 fork E2E tests against live Base chain
- Coverage: worknet lifecycle, DAO governance, merkle claims,
  emission decay, gasless EIP-712, allocation, binding chain

+++ contracts/src/interfaces/IAWPEmission.sol
@@ -38,10 +38,10 @@ interface IAWPEmission {
     function currentEpoch() external view returns (uint256);
     function getRecipientCount() external view returns (uint256);
     function getRecipient(uint256 index) external view returns (address);
-    function getWeight(address addr) external view returns (uint96);
+    function getWeight(address addr) external view returns (uint64);
     function getTotalWeight() external view returns (uint256);
     function getEpochRecipientCount(uint256 epoch) external view returns (uint256);
-    function getEpochWeight(uint256 epoch, address addr) external view returns (uint96);
+    function getEpochWeight(uint256 epoch, address addr) external view returns (uint64);
     function getEpochTotalWeight(uint256 epoch) external view returns (uint256);
 
     // ── Events ──
